Remove conflicting copy of NIO method
In #98, @tachyonics added a vendored copy of a NIO method on
EventLoopFuture to avoid needing to add availability guards. I don't
think that should have been necessary then, but it certainly isn't now,
as smoke-http's minimum deployment targets match the extension in NIO.

This extension contravened one of NIO's requirements on its dependencies:
https://github.com/apple/swift-nio/blob/main/docs/public-api.md#4-extending-nio-types.
Ideally when it was added, it should have been given a different name.

This extension has been working, but it is now conflicting with NIO's main on
Swift 5.9. This patch removes the now-vestigial extension, as it isn't needed.